+++ src/components/calendar/MonthRangePicker/MonthRangePicker.tsx
@@ -323,6 +323,8 @@ export const MonthRangePicker = ({
               isOpen={open}
               size={size}
               pickerOnly={pickerOnly}
+              minDate={minDate}
+              maxDate={maxDate}
               onCalendarClick={() => !disabled && handleOpenChange(!open)}
             />
           </div>